requirement.txt
mysql-connector-python
fastapi
uvicorn
requests
geopy
pandas
mysql-connector-python
pytz

